requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
flatbuffers==23.1.4
gast==0.5.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.4.0
urllib3==1.26.13
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.14.1